__init__ IS the module
